refactor: deleteTodoApi를 따로 파일 분리
